+++ src/core/plugin/index.ts
@@ -1,3 +1,3 @@
 export * from "./StingerloomPlugin";
 export * from "./PluginContext";
-export * from "./mutation";
+export * from "./buffer";