+++ src/core/plugin/mutation/index.ts
@@ -1,5 +1,0 @@
-export * from "./MutationEntry";
-export * from "./MutationPreview";
-export * from "./MutationStrategy";
-export * from "./Mutation";
-export * from "./mutationPlugin";